Move internal addons to their own AddonLoader
